Utilities for crust conversion and Rahul's crust in jose-compatible format
The idea is that it seems that many crust files (or EOS files in general) do not use the units that jose requires for input. Therefore, here are some more lines of code to convert units as well as another crust EOS file but with the correct units. I have checked that the conversion seems to result in the correct orders of magnitude.