More flexible crust support
Users can now specify a filename outside of jose, as before, but also just give a name of a crust that is already on the jose repo and we will fetch the location and load it. Also, users can give a maximal density and have the crust only be returned below that density.